+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-29T11:58:31.816Z",
+    "2026-05-29T13:09:23.492Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {