+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -409,7 +409,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KSSSKVK94VWN12YJ25W4FW68"),
+  _id: z.string().default("01KSSXNJ2M9ZFMGABHVKTJGBQ1"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,