+++ packages/orq-rc/RELEASES.md
@@ -9579,3 +9579,13 @@ Based on:
 - [typescript v4.10.0-rc.56] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.56] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.56 - packages/orq-rc
+
+## 2026-05-29 13:12:38
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.57] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.57] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.57 - packages/orq-rc